splittinbg up emitter/set
