+++ lib/emitter/index.js
@@ -13,7 +13,7 @@ module.exports = new Base({
   inject: [
     require('./storage'),
     require('./off'),
-    require('./on'),
+    require('./set'),
     require('./once'),
     require('./emit')
   ],
