+++ lib/emitter/set/on.js
@@ -1,24 +1,16 @@
 'use strict'
 const vstamp = require('vigour-stamp')
-const Base = require('vigour-base')
-const isObj = require('vigour-util/is/plainobj')
-const isNumberLike = require('vigour-util/is/numberlike')
-const listens = require('./listens')
+const listens = require('../listens')
 const ListensOnattach = listens.Attach
 const ListensOnBase = listens.Base
-const id = require('./id')
-
-// @todo: split this file up -- can be greatly improved
-// @todo: its dirty, its too much code, all slow, unoptmized, double
-// same here make it success
+const id = require('../id')
 
 exports.define = {
   on (val, key, unique, stamp) {
-    // start at zero faster looping!
     if (!key) { key = id(this, val) }
     if (typeof val === 'function') {
       add(this, 'fn', val, key, unique, stamp)
-    } else if (val instanceof Base) {
+    } else if (typeof val === 'object' && val.isBase) {
       if (!add(this, 'base', val, key, unique || true, stamp)) {
         if (!val.listensOnBase) {
           val.setKeyInternal(
@@ -45,7 +37,7 @@ exports.define = {
           val = val.slice(0, 2)
         }
         let passedOnBased = val[1]
-        if (passedOnBased instanceof Base) {
+        if (typeof val === 'object' && val.isBase) {
           if (!passedOnBased.listensOnAttach) {
             passedOnBased
               .setKeyInternal(
@@ -66,33 +58,6 @@ exports.define = {
           listensOnAttach[key] = this
           listensOnAttach.addKey(key, this)
         }
-      }
-    }
-  },
-  set (val, stamp, nocontext) {
-    if (isObj(val) && !(val instanceof Array)) {
-      set(this, val, stamp, nocontext)
-    } else {
-      this.on(val, 'val', void 0, stamp, void 0, nocontext)
-    }
-    return this
-  }
-}
-
-function set (emitter, val, stamp, nocontext) {
-  let props = emitter.properties
-  for (let key in val) {
-    if (props[key]) {
-      props[key].call(emitter, val[key], stamp, nocontext, key)
-    } else {
-      if (!emitter._id && isNumberLike(key)) {
-        // @todo: is this really nessecary? -- maybe better to check in the id generator?
-        emitter._id = Number(key)
-      }
-      if (val[key] === null) {
-        emitter.off(key)
-      } else {
-        emitter.on(val[key], key, void 0, stamp, void 0, nocontext)
       }
     }
   }
